# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mana Ooru - Smart Village Crop Procurement Platform
 * OpenAPI spec version: 0.1.0
 */
export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface OtpSendBody {
  /**
   * @minLength 10
   * @maxLength 15
   */
  mobile: string;
}

export interface OtpSendResponse {
  success: boolean;
  /** Only present in development — the OTP for testing. */
  devOtp?: string;
}

export interface OtpVerifyBody {
  mobile: string;
  /**
   * @minLength 6
   * @maxLength 6
   */
  otp: string;
}

export interface OtpVerifyResponse {
  success: boolean;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface HealthStatus {
  status: string;
}

export type FarmerPaymentStatus =
  (typeof FarmerPaymentStatus)[keyof typeof FarmerPaymentStatus];

export const FarmerPaymentStatus = {
  Pending: "Pending",
  Completed: "Completed",
} as const;

/**
 * @nullable
 */
export type FarmerCropStatus =
  | (typeof FarmerCropStatus)[keyof typeof FarmerCropStatus]
  | null;

export const FarmerCropStatus = {
  On_Hold: "On Hold",
  Partially_Sold: "Partially Sold",
  Fully_Sold: "Fully Sold",
  Sold_Out: "Sold Out",
} as const;

export interface Farmer {
  id: number;
  name: string;
  village: string;
  crop: string;
  quantity: number;
  moisture: string;
  paymentStatus: FarmerPaymentStatus;
  bankAccount: string;
  /** @nullable */
  cropGrade?: string | null;
  /** @nullable */
  harvestDate?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  profilePhotoUrl?: string | null;
  mediaUrls: string[];
  /** @nullable */
  cropStatus?: FarmerCropStatus;
  createdAt: string;
}

export type FarmerInputCropStatus =
  (typeof FarmerInputCropStatus)[keyof typeof FarmerInputCropStatus];

export const FarmerInputCropStatus = {
  On_Hold: "On Hold",
  Partially_Sold: "Partially Sold",
  Fully_Sold: "Fully Sold",
  Sold_Out: "Sold Out",
} as const;

export interface FarmerInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  village: string;
  /** @minLength 1 */
  crop: string;
  /** @minimum 1 */
  quantity: number;
  /** @minLength 1 */
  moisture: string;
  /** @minLength 1 */
  bankAccount: string;
  cropGrade?: string;
  harvestDate?: string;
  notes?: string;
  profilePhotoUrl?: string;
  mediaUrls?: string[];
  cropStatus?: FarmerInputCropStatus;
}

export type FarmerUpdateCropStatus =
  (typeof FarmerUpdateCropStatus)[keyof typeof FarmerUpdateCropStatus];

export const FarmerUpdateCropStatus = {
  On_Hold: "On Hold",
  Partially_Sold: "Partially Sold",
  Fully_Sold: "Fully Sold",
  Sold_Out: "Sold Out",
} as const;

export interface FarmerUpdate {
  name?: string;
  village?: string;
  crop?: string;
  quantity?: number;
  moisture?: string;
  bankAccount?: string;
  cropGrade?: string;
  harvestDate?: string;
  notes?: string;
  profilePhotoUrl?: string;
  mediaUrls?: string[];
  cropStatus?: FarmerUpdateCropStatus;
}

export type PaymentStatusUpdatePaymentStatus =
  (typeof PaymentStatusUpdatePaymentStatus)[keyof typeof PaymentStatusUpdatePaymentStatus];

export const PaymentStatusUpdatePaymentStatus = {
  Pending: "Pending",
  Completed: "Completed",
} as const;

export interface PaymentStatusUpdate {
  paymentStatus: PaymentStatusUpdatePaymentStatus;
}

export interface DashboardSummary {
  totalFarmers: number;
  totalQuantity: number;
  pendingPayments: number;
  completedPayments: number;
  villages: string[];
  crops: string[];
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};