# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mana Ooru - Smart Village Crop Procurement Platform
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: farmers
    description: Farmer procurement records
  - name: dashboard
    description: Dashboard summary and analytics
  - name: Storage
    description: Object storage upload and serving endpoints.
paths:
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect after login.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.

  /otp/send:
    post:
      tags: [Auth]
      operationId: sendOtp
      summary: Send OTP to a mobile number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpSendBody'
      responses:
        '200':
          description: OTP sent successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpSendResponse'
  /otp/verify:
    post:
      tags: [Auth]
      operationId: verifyOtp
      summary: Verify an OTP for a mobile number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpVerifyBody'
      responses:
        '200':
          description: OTP verified successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpVerifyResponse'
        '400':
          description: Invalid or expired OTP.

  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /farmers:
    get:
      operationId: listFarmers
      tags: [farmers]
      summary: List all farmers
      responses:
        "200":
          description: List of farmers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Farmer"
    post:
      operationId: createFarmer
      tags: [farmers]
      summary: Add a new farmer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmerInput"
      responses:
        "201":
          description: Farmer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farmer"
        "400":
          description: Validation error

  /farmers/{id}:
    get:
      operationId: getFarmer
      tags: [farmers]
      summary: Get a farmer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Farmer record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farmer"
        "404":
          description: Farmer not found
    patch:
      operationId: updateFarmer
      tags: [farmers]
      summary: Update farmer details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmerUpdate"
      responses:
        "200":
          description: Updated farmer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farmer"
        "404":
          description: Farmer not found
    delete:
      operationId: deleteFarmer
      tags: [farmers]
      summary: Delete a farmer record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Farmer not found

  /farmers/{id}/payment:
    patch:
      operationId: updatePaymentStatus
      tags: [farmers]
      summary: Update payment status for a farmer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentStatusUpdate"
      responses:
        "200":
          description: Payment status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farmer"
        "404":
          description: Farmer not found

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get procurement summary stats
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /storage/public-objects/{filePath}:
    get:
      tags: [Storage]
      operationId: getPublicObject
      summary: Serve a public asset
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: File not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve a stored object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    OtpSendBody:
      type: object
      required: [mobile]
      properties:
        mobile:
          type: string
          minLength: 10
          maxLength: 15
    OtpSendResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        devOtp:
          type: string
          description: "Only present in development — the OTP for testing."
    OtpVerifyBody:
      type: object
      required: [mobile, otp]
      properties:
        mobile:
          type: string
        otp:
          type: string
          minLength: 6
          maxLength: 6
    OtpVerifyResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Farmer:
      type: object
      required: [id, name, village, crop, quantity, moisture, paymentStatus, bankAccount, createdAt, mediaUrls]
      properties:
        id:
          type: integer
        name:
          type: string
        village:
          type: string
        crop:
          type: string
        quantity:
          type: number
        moisture:
          type: string
        paymentStatus:
          type: string
          enum: [Pending, Completed]
        bankAccount:
          type: string
        cropGrade:
          type: ["string", "null"]
        harvestDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        profilePhotoUrl:
          type: ["string", "null"]
        mediaUrls:
          type: array
          items:
            type: string
        cropStatus:
          type: ["string", "null"]
          enum: ["On Hold", "Partially Sold", "Fully Sold", "Sold Out", null]
        createdAt:
          type: string

    FarmerInput:
      type: object
      required: [name, village, crop, quantity, moisture, bankAccount]
      properties:
        name:
          type: string
          minLength: 1
        village:
          type: string
          minLength: 1
        crop:
          type: string
          minLength: 1
        quantity:
          type: number
          minimum: 1
        moisture:
          type: string
          minLength: 1
        bankAccount:
          type: string
          minLength: 1
        cropGrade:
          type: string
        harvestDate:
          type: string
        notes:
          type: string
        profilePhotoUrl:
          type: string
        mediaUrls:
          type: array
          items:
            type: string
        cropStatus:
          type: string
          enum: ["On Hold", "Partially Sold", "Fully Sold", "Sold Out"]

    FarmerUpdate:
      type: object
      properties:
        name:
          type: string
        village:
          type: string
        crop:
          type: string
        quantity:
          type: number
        moisture:
          type: string
        bankAccount:
          type: string
        cropGrade:
          type: string
        harvestDate:
          type: string
        notes:
          type: string
        profilePhotoUrl:
          type: string
        mediaUrls:
          type: array
          items:
            type: string
        cropStatus:
          type: string
          enum: ["On Hold", "Partially Sold", "Fully Sold", "Sold Out"]

    PaymentStatusUpdate:
      type: object
      required: [paymentStatus]
      properties:
        paymentStatus:
          type: string
          enum: [Pending, Completed]

    DashboardSummary:
      type: object
      required: [totalFarmers, totalQuantity, pendingPayments, completedPayments, villages, crops]
      properties:
        totalFarmers:
          type: integer
        totalQuantity:
          type: number
        pendingPayments:
          type: integer
        completedPayments:
          type: integer
        villages:
          type: array
          items:
            type: string
        crops:
          type: array
          items:
            type: string

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string