# lib/api-zod/src/generated/types/otpSendBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mana Ooru - Smart Village Crop Procurement Platform
 * OpenAPI spec version: 0.1.0
 */

export interface OtpSendBody {
  /**
   * @minLength 10
   * @maxLength 15
   */
  mobile: string;
}