# lib/api-zod/src/generated/types/otpVerifyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mana Ooru - Smart Village Crop Procurement Platform
 * OpenAPI spec version: 0.1.0
 */

export interface OtpVerifyResponse {
  success: boolean;
}